Moved Python binding for copy_usm_ndarray_to_usm_ndarray into dedicated file
Also moved routine for simplification of iteration space into dedicated file
